requirements.txt
matplotlib==3.9.1.post1
mne==1.7.1
numpy==1.26.0
pandas==2.2.2
Pillow==10.4.0
scipy==1.14.0
torch==2.4.0
torcheeg==1.1.2
torchvision==0.19.0